// ui/src/services/cradle/models/FileReferenceWithNote.ts
export interface FileReferenceWithNote {
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    readonly id?: string;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    mimetype?: string | null;
    /**
     * 
     * @type {Array<OptimizedEntryResponse>}
     * @memberof FileReferenceWithNote
     */
    readonly entities?: Array<OptimizedEntryResponse>;
    /**
     * 
     * @type {number}
     * @memberof FileReferenceWithNote
     */
    fileSize?: number | null;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    fileName?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof FileReferenceWithNote
     */
    readonly timestamp?: Date;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    readonly noteId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    md5Hash?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    sha1Hash?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FileReferenceWithNote
     */
    sha256Hash?: string | null;
}